// src/generated/highlighter-lexer.ts
export interface LexError {
  lexeme:      string;
  line:        number;   // base-1
  column:      number;   // base-1
  type:        'Léxico';
  description: string;
}